# Replace bit field manipulation functions with macros