… jerry-core/jrt/jrt-bit-fields.c
… jerry-core/jrt/jrt-bit-fields.h
… jerry-core/rcs/rcs-records.c