JerryScript-DCO-1.0-Signed-off-by: Akos Kiss akiss@inf.u-szeged.hu