@LaszloLango 2016-02-22
I have related work in https://github.com/LaszloLango/jerryscript/commit/1223b994c7970c2a237f9907f4b0a0786bc18d4f
There is much optimization possibility in `ecma_make_*_value` functions too.