@akiss77 2016-02-22
This patch achieves quite good code size reduction on arm targets (2-8K on release builds).

filename | master/text | master/data | pr/text | pr/data | gain/text | gain/data
---------|-------------|-------------|---------|---------|-----------|----------
release.linux-armv7lhf/jerry | 195943 | 52 | 187192 | 52 | 8751 | 0
release.linux-armv7lhf-cp/jerry | 142148 | 52 | 136555 | 52 | 5593 | 0
release.linux-armv7lhf-cp_minimal/jerry | 87957 | 0 | 85396 | 0 | 2561 | 0
release.linux-armv7lhf-mem_stats/jerry | 196539 | 52 | 188297 | 52 | 8242 | 0
release.linux-armv7lhf-mem_stress_test/jerry | 195575 | 52 | 187588 | 52 | 7987 | 0
release.linux-x86_64/jerry | 278590 | 61 | 277340 | 52 | 1250 | 9
release.linux-x86_64-cp/jerry | 210702 | 61 | 210010 | 52 | 692 | 9
release.linux-x86_64-cp_minimal/jerry | 132993 | 0 | 132442 | 0 | 551 | 0
release.linux-x86_64-mem_stats/jerry | 280071 | 61 | 278549 | 52 | 1522 | 9
release.linux-x86_64-mem_stress_test/jerry | 279031 | 61 | 277469 | 52 | 1562 | 9
release.mcu_stm32f3-cp.bin/jerry | 142748 | 72 | 140232 | 72 | 2516 | 0
release.mcu_stm32f3-cp_minimal.bin/jerry | 82436 | 20 | 80912 | 20 | 1524 | 0
release.mcu_stm32f4-cp.bin/jerry | 142748 | 72 | 140232 | 72 | 2516 | 0
release.mcu_stm32f4-cp_minimal.bin/jerry | 82436 | 20 | 80912 | 20 | 1524 | 0
debug.linux-armv7lhf/jerry | 734016 | 60 | 733252 | 60 | 764 | 0
debug.linux-armv7lhf-cp/jerry | 537776 | 60 | 537020 | 60 | 756 | 0
debug.linux-armv7lhf-cp_minimal/jerry | 368272 | 24 | 366996 | 24 | 1276 | 0
debug.linux-armv7lhf-mem_stats/jerry | 737516 | 60 | 736660 | 60 | 856 | 0
debug.linux-armv7lhf-mem_stress_test/jerry | 734064 | 60 | 733300 | 60 | 764 | 0
debug.linux-x86_64/jerry | 817371 | 68 | 816185 | 68 | 1186 | 0
debug.linux-x86_64-cp/jerry | 635223 | 68 | 633962 | 68 | 1261 | 0
debug.linux-x86_64-cp_minimal/jerry | 443502 | 32 | 442300 | 32 | 1202 | 0
debug.linux-x86_64-mem_stats/jerry | 821176 | 68 | 820710 | 68 | 466 | 0
debug.linux-x86_64-mem_stress_test/jerry | 817391 | 68 | 816205 | 68 | 1186 | 0
debug.mcu_stm32f4-cp.bin/jerry | 547212 | 76 | 546324 | 76 | 888 | 0
debug.mcu_stm32f4-cp_minimal.bin/jerry | 367396 | 44 | 366500 | 44 | 896 | 0

Performance on x86_64/linux also improved slightly, especially on the longer running test cases (crypto-md5.js and crypto-sha1.js).

```
                  Benchmark |                   Perf(+ is better)
                 3d-cube.js |   0.259s ->   0.262s :  -1.093%  (+-1.765%) : [≈]
     access-binary-trees.js |   0.137s ->   0.132s :  +3.519%  (+-1.882%) : [+]
         access-fannkuch.js |   0.719s ->   0.717s :  +0.255%  (+-1.278%) : [≈]
            access-nbody.js |   0.316s ->   0.307s :  +2.847%  (+-0.907%) : [+]
bitops-3bit-bits-in-byte.js |   0.205s ->   0.203s :  +0.893%  (+-2.276%) : [≈]
     bitops-bits-in-byte.js |   0.262s ->   0.269s :  -2.610%  (+-2.358%) : [-]
      bitops-bitwise-and.js |   0.308s ->   0.296s :  +3.794%  (+-1.580%) : [+]
   controlflow-recursive.js |   0.090s ->   0.090s :  -0.557%  (+-7.179%) : [≈]
              crypto-aes.js |   0.508s ->   0.494s :  +2.854%  (+-1.115%) : [+]
              crypto-md5.js |   3.048s ->   2.736s : +10.226%  (+-0.374%) : [+]
             crypto-sha1.js |   1.351s ->   1.213s : +10.178%  (+-0.343%) : [+]
       date-format-tofte.js |   0.214s ->   0.208s :  +2.802%  (+-1.420%) : [+]
       date-format-xparb.js |   0.100s ->   0.097s :  +2.663%  (+-4.337%) : [≈]
             math-cordic.js |   0.293s ->   0.293s :  -0.057%  (+-2.025%) : [≈]
       math-partial-sums.js |   0.160s ->   0.159s :  +1.143%  (+-3.549%) : [≈]
      math-spectral-norm.js |   0.148s ->   0.148s :  +0.337%  (+-2.330%) : [≈]
            string-fasta.js |   0.449s ->   0.430s :  +4.306%  (+-1.561%) : [+]
            Geometric mean: |                 Speed up: 2.501% (+-0.650%) : [+]
```