@akiss77 2016-02-23
Rebased and pushed. (Also: found a solution to get rid of not-so-nice type casts that were present in previous versions of this patch.)