@akiss77 2016-02-23
@zherczeg I've updated the patch as suggested. Does it look better?