@zherczeg 2016-02-25
In general functions are easier to debug but I have no strong opinion, because I also use a lot of macros. I did not find the RPi2 results, am I missing something?