@bzsolt 2016-02-25
RPi2 results
master (https://github.com/Samsung/jerryscript/commit/76b37f34d0df7fadbecf57ac52345c6b50631b54)

                      Benchmark |                  Perf(+ is better)
                     3d-cube.js |  2.627s ->   2.543s :  +3.173%  (+-3.506%) : [≈]
         access-binary-trees.js |  1.407s ->   1.370s :  +2.607%  (+-2.291%) : [+]
             access-fannkuch.js |  7.960s ->   7.967s :  -0.084%  (+-2.530%) : [≈]
                access-nbody.js |  3.070s ->   3.077s :  -0.217%  (+-1.078%) : [≈]
    bitops-3bit-bits-in-byte.js |  1.730s ->   1.723s :  +0.386%  (+-1.912%) : [≈]
         bitops-bits-in-byte.js |  2.653s ->   2.603s :  +1.884%  (+-3.494%) : [≈]
          bitops-bitwise-and.js |  3.680s ->   3.663s :  +0.453%  (+-3.587%) : [≈]
       controlflow-recursive.js |  0.890s ->   0.870s :  +2.247%  (+-0.000%) : [+]
                  crypto-aes.js |  4.730s ->   4.643s :  +1.832%  (+-2.769%) : [≈]
                  crypto-md5.js | 21.490s ->  21.847s :  -1.660%  (+-0.982%) : [-]
                 crypto-sha1.js |  9.750s ->   9.817s :  -0.684%  (+-3.406%) : [≈]
           date-format-tofte.js |  2.233s ->   2.223s :  +0.448%  (+-2.090%) : [≈]
           date-format-xparb.js |  1.087s ->   1.053s :  +3.068%  (+-4.240%) : [≈]
                 math-cordic.js |  2.977s ->   2.943s :  +1.120%  (+-2.463%) : [≈]
           math-partial-sums.js |  1.660s ->   1.670s :  -0.602%  (+-0.000%) : [-]
          math-spectral-norm.js |  1.503s ->   1.457s :  +3.104%  (+-3.064%) : [+]
                string-fasta.js |  4.537s ->   4.467s :  +1.543%  (+-3.251%) : [≈]
                Geometric mean: |                Speed up: 1.106% (+-0.661%) : [+]

Binary sizes (bytes): 197,240 -> 193,144